requirements.txt
jsonlines==2.0.0
numpy==1.19.5
pandas==1.1.5
scikit_learn==0.24.2
torch==1.7.1
tqdm==4.64.0
transformers==4.18.0
Pillow==8.4.0
